DeepseekOcrApi/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
torch>=2.0.0
transformers>=4.35.0
vllm>=0.2.0
Pillow>=10.0.0
PyMuPDF>=1.23.0
img2pdf==0.5.1
tqdm>=4.66.0
numpy>=1.24.0
